requirements.txt
playwright>=1.40
yt-dlp>=2024.1
pyyaml>=6.0
